fixed tweets
Added escape characters (/) in front of commas which were fucking up the html when loaded in.
Removed all the excess columns (such as likes counts) so that only relevant info is pulled.
Fixed the js and css files so that the tweet looks better, does not display like counts, and properly displays the icons.